Split window quality modules
